Use a modin conftest instead.
